requirements.txt
diffusers 
accelerate
transformers
scipy 
ftfy
Pillow
torch
torchvision
torchaudio
jupyter    